feat(dir): cache edited files in the nvim cache (10m)
Before editing the file from the ud fzf picker, cache the file in the
nvim cache. The nvim cache currently stores 1 file only. Possible
improvements in future are to store more than 1 file and display the
cache for the user to select. Even better, sort the files displayed by
frequency of use.